pyproject.toml
[project]
name = "ubipro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "google-genai>=1.59.0",
    "pypdf>=6.6.0",
    "streamlit>=1.53.0",
    "tiktoken>=0.12.0",
]
